feat exceptions/likes/friends

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -3,12 +3,11 @@ package ru.yandex.practicum.filmorate.model;
 import lombok.Data;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 
 @Data
 public class Film {
 
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -20,14 +19,14 @@ public class Film {
 
     private int score = 0;
 
-    /*
+
     public Film(String name, String description, LocalDate releaseDate, int duration) {
         this.name = name;
         this.description = description;
         this.releaseDate = releaseDate;
         this.duration = duration;
     }
-
+    /*
     public Film(String name, String description, LocalDate releaseDate, int duration, ArrayList<Integer> isLiked) {
         this.name = name;
         this.description = description;
